@@ -26,8 +26,8 @@ import android.preference.Preference;
 import android.preference.PreferenceGroup;
 import android.preference.PreferenceManager;
 import android.preference.PreferenceScreen;
-import android.support.v4.app.Fragment;
-import com.zoromatic.timetable.R;
+import androidx.fragment.app.Fragment;
+
 import android.util.TypedValue;
 import android.view.KeyEvent;
 import android.view.LayoutInflater;
